requirements.txt
fastapi
sqlalchemy
psycopg2
uvicorn
pydantic
passlib[bcrypt]
requests
bcrypt==4.0.1
passlib
pyjwt
